requirements.txt
python-dotenv==1.0.0
pypdf==3.15.5
transformers==4.33.1
langchain==0.0.285
docx2txt==0.8
wikipedia==1.4.0
pinecone-client==2.2.2
sentence-transformers==2.2.2